backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# AI/ML
openai==1.10.0
sentence-transformers==2.3.1
google-generativeai==0.3.2
langchain-core>=0.3.67
langgraph>=0.6.0

# LandingAI - Official library for Agentic Document Extraction
landingai-ade>=0.1.0
# Note: agentic-doc is legacy, using landingai-ade instead

# Document processing
PyPDF2==3.0.1
reportlab==4.0.9
pypdf==3.17.4
pdf2image==1.16.3

# Utilities
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings==2.1.0
aiofiles==23.2.1

# Scientific
numpy==1.26.3
scikit-learn==1.4.0
